support reporting files
updated Readme.md file
