+++ src/main/java/com/pages/YourdataPage.java
@@ -2,6 +2,7 @@ package com.pages;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
@@ -38,18 +39,25 @@ public class YourdataPage {
 		return utils.waitForElementPresent( driver,toOffer,40);
 	}
 
-	public boolean verify_default_data(List<String> dataGiven)
+	public boolean verify_default_data(List<Map<String, String>> credList)
 	{	
 		getPageHeader();
+		List<String> dataoptions = new ArrayList<String>();
+		dataoptions.add(credList.get(0).get("Familienstatus"));
+		dataoptions.add(credList.get(0).get("im öffentlichen Dienst beschäftigt"));
+		dataoptions.add(credList.get(0).get("in Rente"));
+		dataoptions.add(credList.get(0).get("Selbstbeteiligung"));
+		
+		
 		List<String> actualdata = new ArrayList<String>();
-		String[] dataHeader = {"Familienstatus", "im öffentlichen Dienst beschäftigt", "in Rente","Selbstbeteiligung"};
+		String[] dataHeader = dataoptions.toArray(new String[0]);
 		for (String strHeader: dataHeader)
 		{
 			By locatorData = By.xpath(strData.replaceAll("data", strHeader));
 			actualdata.add(utils.waitForElementPresent(driver, locatorData).getText());			
 		}
 		
-		if(actualdata.containsAll(dataGiven))
+		if(actualdata.containsAll(dataoptions))
 			return true;
 		else
 			return false;
@@ -58,11 +66,11 @@ public class YourdataPage {
 	
 	
 	
-	public RecommendedPlan click_showtop()
+	public RecommendedPlanPage click_showtop()
 	{	
 		getButtonShowtop().click();
 		getToOffer().click();
-		return new RecommendedPlan(driver);
+		return new RecommendedPlanPage(driver);
 	}
 	
 }
