--- src/test/java/testrunners/ParallelRun.java
@@ -1,28 +1,0 @@
-package testrunners;
-
-import org.junit.runner.RunWith;
-import org.testng.annotations.DataProvider;
-
-import io.cucumber.junit.Cucumber;
-import io.cucumber.testng.AbstractTestNGCucumberTests;
-import io.cucumber.junit.CucumberOptions;
-@RunWith(Cucumber.class)
-@CucumberOptions(
-		plugin = {"pretty","html:target/cucumber-reports/report.html",
-	              "json:target/cucumber.json",
-	              "rerun:target/rerun.txt",
-				"com.aventstack.extentreports.cucumber.adapter.ExtentCucumberAdapter:",
-				"timeline:test-output-thread/"
-				}, 
-		monochrome = true,
-		features = {"src/test/resources/com/features"},
-		glue = {"stepdefinitions", "hooks"}
-)
-
-public class ParallelRun extends AbstractTestNGCucumberTests {
-	@Override
-	@DataProvider(parallel = true)
-	public Object[][] scenarios() {
-		return super.scenarios();
-	}
-}
