+++ src/test/resources/com/properties/BasePage.properties
@@ -1,0 +1,3 @@
+
+
+cookiepopup=css$button[class='btn btn-primary cookies-banner__opt-in']
