+++ src/test/resources/com/properties/Contracts.properties
@@ -1,6 +1,5 @@
 
 
-cookiepopup=css$button[class='btn btn-primary cookies-banner__opt-in']
 contractpage=xpath$//ul[@class='_segment_15ir2n _sections_1rlfr8']/li/a[text()='Verträge']
 menulocators=xpath$//ul[@class='_segment_15ir2n _sections_1rlfr8']/li/a
 str_angebote=//ul[@class='_segment_15ir2n _sections_1rlfr8']/li/a[text()='deal_option']
