+++ src/test/resources/com/properties/Verify_contract.properties
@@ -3,4 +3,4 @@
 
 popup=xpath$//span[@class='_block_dsfphm' and contains(.,'Nein')]
 close_pop_up=xpath$//button[starts-with(@class,'cucumber-close-modal')]
-offer_verify=xpath$//div[@class = '_title_niboal' and contains(.,'Privathaftpflicht')]
+offer_verify=//div[@class = '_title_niboal' and contains(.,'name')]
