+++ src/test/resources/extent-config.xml
@@ -1,0 +1,44 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<extentreports>
+	<configuration>
+		<!-- report theme -->
+
+		<!-- standard, dark -->
+
+		<theme>dark</theme>
+		<!-- document encoding -->
+
+		<!-- defaults to UTF-8 -->
+
+		<encoding>UTF-8</encoding>
+		<!-- protocol for script and stylesheets -->
+
+		<!-- defaults to https -->
+
+		<protocol>http</protocol>
+		<!-- title of the document -->
+		<documentTitle>Extent</documentTitle>
+		<!-- report name - displayed at top-nav -->
+
+		<reportName>Grasshopper Report</reportName>
+		<!-- location of charts in the test view -->
+
+		<!-- top, bottom -->
+
+		<testViewChartLocation>bottom</testViewChartLocation>
+		<!-- custom javascript -->
+
+		<scripts>
+
+<![CDATA[
+$(document).ready(function() {
+});
+]]>
+		</scripts>
+		<!-- custom styles -->
+		<styles>
+<![CDATA[
+]]>
+		</styles>
+	</configuration>
+</extentreports>
