+++ src/test/resources/extent.properties
@@ -1,0 +1,17 @@
+extent.reporter.spark.start=true
+extent.reporter.spark.out=test-output/SparkReport/Spark.html
+extent.reporter.spark.config=./src/test/resources/extent-config.xml
+
+extent.reporter.spark.out=test-output/SparkReport/
+
+screenshot.dir=test-output/
+screenshot.rel.path=../Screenshots/
+extent.reporter.pdf.start=true
+extent.reporter.pdf.out=test-output/PdfReport/ExtentPdf.pdf
+basefolder.name=test-output/SparkReport
+basefolder.datetimepattern=d-MMM-YY HH-mm-ss
+extent.reporter.spark.vieworder=dashboard,test,category,exception,author,device,log
+systeminfo.os=Windows
+systeminfo.user=Saurabh
+systeminfo.build=1.1
+systeminfo.AppName=AutomationPractice
